Backend/requirements.txt
Flask
pillow
numpy
tensorflow==1.14
tflearn
click
flask-mysqldb